[repositoryminer-technicaldebt] - bug fix
This commit solves a bug that occurred when the user tried to get all
indicators by their types and pass them to the TD analysis filter. In
order to solve this problem, I removed some methods and added other new
ones, with the older architecture was not possible solve this problem.