[rm-findbugs] Solved bug in the analysis.
When the user passes a jar that is dependency is possible not to find
the source code associated with the binary inside the project. This
situation caused the analysis stop, due the source files not found. This
situation was solved just for considering only a binary if its
correspondent source code is in the project.